Split `TextChunk` to separate file
